+++ tests/Unit/PersonioIntegration/Api_Request.php
@@ -29,15 +29,15 @@ class Api_Request extends PersonioTestCase {
 	}
 
 	/**
-	 * Test an invalid API request.
+	 * Test a valid API request.
 	 *
 	 * @return void
 	 */
-	public function test_invalid_request(): void {
+	public function test_valid_request(): void {
 		// create the request.
 		$request_object = new \PersonioIntegrationLight\PersonioIntegration\Api_Request();
 		$request_object->set_url( 'https://api.personio.de/v2/auth/token' );
-		$request_object->set_post_data( array( 'key' => 'value' ) );
+		$request_object->set_post_data( array( 'client_id' => 'valid' ) );
 
 		// send it.
 		$request_object->send();
@@ -49,6 +49,30 @@ class Api_Request extends PersonioTestCase {
 		$this->assertJson( $response );
 
 		// check if the response contains the key "access_token".
+		$this->assertStringContainsString( '"access_token"', $response );
+	}
+
+	/**
+	 * Test an invalid API request.
+	 *
+	 * @return void
+	 */
+	public function test_invalid_request(): void {
+		// create the request.
+		$request_object = new \PersonioIntegrationLight\PersonioIntegration\Api_Request();
+		$request_object->set_url( 'https://api.personio.de/v2/auth/token' );
+		$request_object->set_post_data( array( 'client_id' => 'invalid' ) );
+
+		// send it.
+		$request_object->send();
+
+		// get the response.
+		$response = $request_object->get_response();
+		$this->assertNotEmpty( $response );
+		$this->assertIsString( $response );
+		$this->assertJson( $response );
+
+		// check if the response contains the key "invalid_request".
 		$this->assertStringContainsString( '"invalid_request"', $response );
 	}