+++ changelog.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Added multiple new PHP Unit Tests
+- Added some new hooks
 - Added support for detection of ACF Pro
 
 ### Changed
@@ -16,6 +17,7 @@
 ### Fixed
 
 - Fixed a missing check for direct access in two templates
+- Fixed to show automatic import in log for deletion of a single position
 
 ## [5.1.1] - 05.01.2026